requirements.txt
Flask==2.2.2
flask_cors
twitter-api-client==0.7.9
